# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kala API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/lessons/{course_id}/skill/{skill_id}": {
      "get": {
        "tags": [
          "lessons"
        ],
        "summary": "Get Lesson",
        "description": "Load the skill's guided lesson, generating and persisting it on first\nrequest. Idempotent thereafter.",
        "operationId": "get_lesson_lessons__course_id__skill__skill_id__get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/lessons/{course_id}/steps/{step_id}/check": {
      "post": {
        "tags": [
          "lessons"
        ],
        "summary": "Submit Check",
        "description": "Grade a step's comprehension check server-side. The step advances (the\nread -> understand -> apply gate) only when the client sees correct=True.\nA pass writes a tutor evidence_event and moves the tracer; the schedule is\nseeded too so a checked concept enters spaced review like any other card.\n\nOwnership is verified in two independent steps before anything is\nwritten, because the URL's course_id is caller-supplied and must not be\ntrusted on its own:\n  1. the step -> its lesson_id (a step id + check_item_id is not, by\n     itself, proof the step belongs to THIS course);\n  2. that lesson_id -> its lesson row, filtered by course_id AND\n     institution_id from the path. Only then is the step known to\n     actually belong here, closing the hole where a client could submit\n     /lessons/COURSE_B/steps/{step-from-course-A}/check and have\n     evidence + mastery written against COURSE_B using a check item that\n     was never generated for it.",
        "operationId": "submit_check_lessons__course_id__steps__step_id__check_post",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "step_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Step Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckBody": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "choice_id": {
            "type": "string",
            "title": "Choice Id"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms",
            "default": 0
          },
          "hints_used": {
            "type": "integer",
            "title": "Hints Used",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "choice_id"
        ],
        "title": "CheckBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}